requirements.txt
pillow>=10.0.0
aiohttp>=3.8.0
